add DBF read support, combining shapes and metadata in Shapefile.Table (#33)
* add DBF read support, combining shapes and metadata in Shapefile.Table

Shapefile.Table implements the Tables.jl interface.

Going forward it would be nice to use this as a basis to try to develop a kind of GeoTables interface
like mentioned in the README of https://github.com/visr/GeoJSONTables.jl. With GeoJSONTables and Shapefile we will then have two native formats that already implement the Tables interface, that we can use for developing GeoTables.

Depends on https://github.com/JuliaData/DBFTables.jl/pull/9

Closes #19

Note that this does not make use of the shx index files, but since we read all shapes into a vector, there is no need for this, since we can already easily find specific shapes.